+++ packages/workflow-core/src/lib/plugins/external-plugin/api-plugin/constants.ts
@@ -1,0 +1,1 @@
+export const REQUEST_PAYLOAD_BLACKLIST = ['callbackUrl'] as const;
