+++ packages/workflow-core/src/lib/plugins/external-plugin/api-plugin/index.ts
@@ -1,0 +1,1 @@
+export * from './api-plugin';
